refactor: 이름 변경 및 폴더구조 변경 useGetProduct -> getProduct
